+++ TombEngine/Game/effects/tomb4fx.cpp
@@ -381,7 +381,7 @@ void UpdateFireProgress()
 
 void AddFire(int x, int y, int z, short roomNum, float size, short fade)
 {
-	AddFire(Vector3i(x, y, z), roomNum, Vector4::One, size, fade);
+	AddFire(Vector3i(x, y, z), roomNum, NEUTRAL_COLOR, size, fade);
 }
 
 void AddFire(Vector3i& pos, int roomNumber, Vector4 color, float size, short fade)