+++ CHANGELOG.md
@@ -7,8 +7,10 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 
 ### New features
 * Added ease-in and ease-out to flyby camera movement when the "Freeze camera" flag is set.
-
+* Added gamma correction setting.
+ 
 ### Bug fixes
+* Fixed incorrect dynamic range for vertex colors, ambient light, dynamic lights and particle effects.
 * Fixed flyby camera jitter by converting the spline type to floating-point.
 
 ### Lua API changes